requirements.txt
chromadb
google-generativeai
gradio
openai
jupyter
tiktoken
